Add citylight theme
Even though I can't figure out where I got this
file from... It was somewhere on GitHub, I should
have written the link into it. It is not exactly
identical to the kitty theme with the same name,
so I'd like to fix that. But this will do for now.